refactor(docs): move support matrix schema in external module
